Adding individual package configs along with saving.
The tangled mess that is the set of default folders, parsers, and other configs for simply the dns package is getting in the way of clean understandable code and extensability that will be required for any community extension. I am pushing things off into a package system to more clearly seperate what code belongs to what package.