Added default copilot.local DNS redirection.
ISP's like verizon, much like this plugin, do DNS-redirection. They
often do it for advertising, in fact, they do it at my apartment. This
has finaly been useful as it helped my ID that when ad based DNS
redirection is occuring the copilot.local webpage will not be
accessible. The solution seems to be to have the dns redirection on by
default for the copilot.local page. Ironically, copilot.local DNS
redirection is needed by default to combat ISP redirection.