Added logo to blockpage.
The blockpage was images of the grafitti from Turkey during
DNS censoring. This required the trainer to know the history
of that specific incident to explain the blockpage. This is
an anti-pattern as it is an undocumented requirement and
it imposes a narrative facilitation style onto the trainer.
As such, they have been removed and replaced with the logo.